+++ CHANGELOG.md
@@ -6,4 +6,7 @@
 * * Added the `spawn_item_particles` parameter. If false, prevents item particles from being spawned. True by default.
 * * Added the `cooldown` parameter. Allows you to set a custom cooldown.
 * * Added `override_vanilla` parameter. If true, prevents **ALL** vanilla behaviors from being executed. This should never be used on block items, as it will make the block unplaceable.
+* New Tweak! `/damage` backport. Not exact copy of the command, but still pretty useful.
+* m-tweaks now collects analytic data, obviously you can turn it off. Data sent: 
+* * Mod version, Minecraft version, OS, Java version, and current config.
 * Updated Chinese translation. Courtesy of [Rad233](https://github.com/Rad233).